refactor: add firebase in separate file
