+++ src/index.tsx
@@ -1,23 +1,6 @@
 import ReactDOM from "react-dom/client";
-
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
 import "./index.scss";
 import App from "./App";
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_API_KEY,
-  authDomain: "genial-motif-376109.firebaseapp.com",
-  projectId: "genial-motif-376109",
-  storageBucket: "genial-motif-376109.appspot.com",
-  messagingSenderId: "100913802659",
-  appId: "1:100913802659:web:568843dcad439cca623e16",
-  measurementId: "G-GFD0HL1NJL",
-};
-
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-console.log(analytics);
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
